refactor(adm/plots): split Plots.py into a package
The 2167-line python/pdstools/adm/Plots.py is decomposed into a
python/pdstools/adm/Plots/ package, with methods grouped by concern as
mixins composed by a thin Plots class in __init__.py.

Layout
------
- _helpers.py    – requires decorator, MetricLimits annotation helper,
                   add_bottom_left_text_to_bubble_plot, distribution_graph,
                   COLORSCALE_TYPES.
- _base.py       – _PlotsBase shared attribute surface (datamart) used by
                   every mixin.
- _overview.py   – bubble_chart, tree_map, action_overlap,
                   partitioned_plot.
- _performance.py – over_time, gains_chart,
                    performance_volume_distribution,
                    proposition_success_rates.
- _score.py      – score_distribution, multiple_score_distributions.
- _predictors.py – predictor_performance, predictor_category_performance,
                   predictor_contribution, predictor_performance_heatmap,
                   predictor_count, _boxplot_pre_aggregated helper.
- _binning.py    – predictor_binning, multiple_predictor_binning,
                   binning_lift.
- __init__.py    – composes the mixins into the public Plots class and
                   re-exports fig_update_facet, add_metric_limit_lines,
                   add_bottom_left_text_to_bubble_plot, distribution_graph,
                   requires, COLORSCALE_TYPES so existing imports such as
                   'from pdstools.adm.Plots import Plots' and
                   'from pdstools.adm.Plots import add_metric_limit_lines'
                   continue to resolve.

Public surface unchanged
------------------------
sorted([n for n in dir(Plots) if not n.startswith('_')]) is identical
before and after this refactor; method bodies are byte-for-byte equivalent
(only the helper imports moved from same-module to ._helpers /
...utils.plot_utils). The full pytest suite (1690 passed, 11 skipped)
runs green and pre-commit passes clean.

This is a pure code-organisation change in service of the v5 release;
no behavioural changes.
