--- python/pdstools/adm/Plots.py
@@ -1,2167 +1,0 @@
-from __future__ import annotations
-
-__all__ = ["Plots", "fig_update_facet"]
-import logging
-from collections.abc import Iterable
-from datetime import timedelta
-from functools import wraps
-from typing import (
-    TYPE_CHECKING,
-    Literal,
-    TypeVar,
-    overload,
-)
-from collections.abc import Callable
-
-import polars as pl
-import polars.selectors as cs
-from typing_extensions import ParamSpec
-from typing import Concatenate
-
-from ..utils import cdh_utils
-from ..utils.metric_limits import MetricLimits
-from ..utils.namespaces import LazyNamespace
-from ..utils.plot_utils import (
-    LIFT_DIRECTION_COLORS,
-    Figure,
-    abbreviate_label_expr,
-    fig_update_facet,
-    get_colorscale,
-    simplify_facet_titles,
-)
-from ..utils.types import QUERY
-
-__all__ = ["Plots", "fig_update_facet"]
-
-logger = logging.getLogger(__name__)
-
-if TYPE_CHECKING:  # pragma: no cover
-    from .ADMDatamart import ADMDatamart
-COLORSCALE_TYPES = list[tuple[float, str]] | list[str]
-
-T = TypeVar("T", bound="Plots")
-P = ParamSpec("P")
-
-
-def requires(
-    model_columns: Iterable[str] | None = None,
-    predictor_columns: Iterable[str] | None = None,
-    combined_columns: Iterable[str] | None = None,
-):
-    def decorator(
-        func: Callable[Concatenate[T, P], Figure | pl.LazyFrame],
-    ) -> Callable[Concatenate[T, P], Figure | pl.LazyFrame]:
-        @overload
-        def wrapper(  # type: ignore[valid-type]  # mypy does not support P.args in @overload stubs
-            self: T,
-            *args: P.args,
-            return_df: Literal[False] = ...,
-            **kwargs: P.kwargs,
-        ) -> Figure: ...
-
-        @overload
-        def wrapper(  # type: ignore[valid-type]  # mypy does not support P.args in @overload stubs
-            self: T,
-            *args: P.args,
-            return_df: Literal[True],
-            **kwargs: P.kwargs,
-        ) -> pl.LazyFrame: ...
-
-        @wraps(func)
-        def wrapper(  # type: ignore[valid-type]  # mypy does not support P.args in @overload stubs
-            self: T,
-            *args: P.args,
-            return_df: bool = False,
-            **kwargs: P.kwargs,
-        ) -> Figure | pl.LazyFrame:
-            # Validation logic (unchanged)
-            if model_columns:
-                if self.datamart.model_data is None:
-                    raise ValueError("Missing data: model_data")
-                missing = {c for c in model_columns if c not in self.datamart.model_data.collect_schema().names()}
-                if missing:
-                    raise ValueError(f"Missing required model columns:{missing}")
-
-            if predictor_columns:
-                if self.datamart.predictor_data is None:
-                    raise ValueError("Missing data: predictor_data")
-                missing = {
-                    c for c in predictor_columns if c not in self.datamart.predictor_data.collect_schema().names()
-                }
-                if missing:
-                    raise ValueError(f"Missing required predictor columns:{missing}")
-
-            if combined_columns:
-                if self.datamart.combined_data is None:
-                    raise ValueError("Missing data: combined_data")
-                missing = {c for c in combined_columns if c not in self.datamart.combined_data.collect_schema().names()}
-                if missing:
-                    raise ValueError(f"Missing required combined columns:{missing}")
-
-            return func(self, *args, return_df=return_df, **kwargs)  # type: ignore[arg-type]  # ParamSpec forwarding: mypy loses track of kwargs type through *args/**kwargs
-
-        return wrapper
-
-    return decorator
-
-
-def add_bottom_left_text_to_bubble_plot(
-    fig: Figure,
-    df: pl.LazyFrame,
-    bubble_size: int,
-):
-    def get_nonperforming_models(df: pl.LazyFrame):
-        return (
-            df.filter(
-                (pl.col("Performance") == 50) & ((pl.col("SuccessRate").is_null()) | (pl.col("SuccessRate") == 0)),
-            )
-            .select(pl.first().count())
-            .collect()
-            .item()
-        )
-
-    if len(fig.layout.annotations) > 0:
-        for i in range(len(fig.layout.annotations)):
-            oldtext = fig.layout.annotations[i].text.split("=")
-            subset = df.filter(pl.col(oldtext[0]) == oldtext[1])
-            num_models = subset.select(pl.first().count()).collect().item()
-            if num_models > 0:
-                bottomleft = get_nonperforming_models(subset)
-                newtext = f"{num_models} models: {bottomleft} ({round(bottomleft / num_models * 100, 2)}%) at (50,0)"
-                fig.layout.annotations[i].text += f"<br><sup>{newtext}</sup>"
-                if len(fig.data) > i:
-                    fig.data[i].marker.size *= bubble_size
-                else:
-                    print(fig.data, i)
-        return fig
-    num_models = df.select(pl.first().len()).collect().item()
-    bottomleft = get_nonperforming_models(df)
-    newtext = f"{num_models} models: {bottomleft} ({round(bottomleft / num_models * 100, 2)}%) at (50,0)"
-    fig.layout.title.text += f"<br><sup>{newtext}</sup>"
-    fig.data[0].marker.size *= bubble_size
-    return fig
-
-
-def add_metric_limit_lines(
-    fig: Figure,
-    metric_id: str = "ModelPerformance",
-    scale: float = 100.0,
-    orientation: str = "vertical",
-) -> Figure:
-    """Add dashed lines at MetricLimits thresholds (red=hard, green=best practice)."""
-    limits = MetricLimits.get_limit_for_metric(metric_id)
-    if not limits:
-        return fig
-
-    add_line = fig.add_vline if orientation == "vertical" else fig.add_hline
-    pos_key = "x" if orientation == "vertical" else "y"
-
-    line_specs = [
-        ("minimum", "rgba(255, 69, 0, 0.5)", "Min"),
-        ("best_practice_min", "rgba(0, 128, 0, 0.5)", "Good"),
-        ("best_practice_max", "rgba(0, 128, 0, 0.5)", "Good"),
-        ("maximum", "rgba(255, 69, 0, 0.5)", "Max"),
-    ]
-
-    for limit_key, color, label in line_specs:
-        value = limits.get(limit_key)
-        if value is not None:
-            scaled = value * scale
-            add_line(
-                **{pos_key: scaled},
-                line_dash="dash",
-                line_width=1,
-                line_color=color,
-                layer="below",
-                annotation_text=f"{label} ({scaled:.0f})",
-                annotation_position="top" if orientation == "vertical" else "right",
-                annotation_font_size=9,
-                annotation_font_color=color,
-            )
-
-    return fig
-
-
-def distribution_graph(df: pl.LazyFrame, title: str):
-    import plotly.graph_objects as go
-    from plotly.subplots import make_subplots
-
-    plot_df = df.collect()
-    fig = make_subplots(specs=[[{"secondary_y": True}]])
-    fig.add_trace(
-        go.Bar(x=plot_df["BinSymbol"], y=plot_df["BinResponseCount"], name="Responses"),
-    )
-    fig.add_trace(
-        go.Scatter(
-            x=plot_df["BinSymbol"],
-            y=plot_df["BinPropensity"],
-            yaxis="y2",
-            name="Propensity",
-            mode="lines+markers",
-        ),
-    )
-    fig.update_layout(
-        template="pega",
-        title=title,
-        xaxis_title="Range",
-        yaxis_title="Responses",
-    )
-    fig.update_yaxes(title_text="Propensity", secondary_y=True)
-    fig.layout.yaxis2.tickformat = ",.3%"
-    fig.layout.yaxis2.zeroline = False
-    fig.update_yaxes(showgrid=False)
-    fig.update_xaxes(type="category")
-
-    return fig
-
-
-class Plots(LazyNamespace):
-    dependencies = ["plotly"]
-    dependency_group = "adm"
-
-    def __init__(self, datamart: "ADMDatamart"):
-        self.datamart = datamart
-        super().__init__()
-
-    @requires({"ModelID", "Performance", "SuccessRate", "ResponseCount", "Name"})
-    def bubble_chart(
-        self,
-        *,
-        last: bool = True,
-        rounding: int = 5,
-        query: QUERY | None = None,
-        facet: str | pl.Expr | None = None,
-        color: str | None = "Performance",
-        show_metric_limits: bool = False,
-        return_df: bool = False,
-    ):
-        """The Bubble Chart, as seen in Prediction Studio
-
-        Parameters
-        ----------
-        last : bool, optional
-            Whether to only include the latest snapshot, by default True
-        rounding: int, optional
-            To how many digits to round the performance number
-        query : Optional[QUERY], optional
-            The query to apply to the data, by default None
-        facet : Optional[Union[str, pl.Expr]], optional
-            Column name or Polars expression to facet the plot into subplots, by default None
-        show_metric_limits : bool, optional
-            Whether to show dashed vertical lines at the ModelPerformance
-            metric limit thresholds (from MetricLimits.csv), by default False
-        return_df : bool, optional
-            Whether to return a dataframe instead of a plot, by default False
-
-        Returns
-        -------
-        Figure | pl.LazyFrame
-            Plotly scatter figure or LazyFrame if return_df=True
-
-        Examples
-        --------
-        >>> from pdstools import ADMDatamart
-        >>> dm = ADMDatamart.from_ds_export(base_path="/my_export_folder")
-
-        >>> # Basic bubble chart using the latest snapshot
-        >>> fig = dm.plot.bubble_chart()
-
-        >>> # Facet by Channel to see performance per channel
-        >>> fig = dm.plot.bubble_chart(facet="Channel")
-
-        >>> # Include all snapshots and show metric-limit guidelines
-        >>> fig = dm.plot.bubble_chart(last=False, show_metric_limits=True)
-
-        >>> # Return the underlying data instead of the figure
-        >>> df = dm.plot.bubble_chart(return_df=True)
-
-        """
-        # why do we need this select? it's not ideal because columns used in the query are not always selected
-        columns_to_select = list(
-            set(
-                [
-                    "ModelID",
-                    color,
-                    "SuccessRate",
-                    "ResponseCount",
-                    "ModelTechnique",
-                    "SnapshotTime",
-                    "LastUpdate",
-                    "Configuration",
-                    *self.datamart.context_keys,
-                ],
-            ),
-        ) + (["Performance"] if color != "Performance" else [])
-        if facet is not None:
-            if isinstance(facet, pl.Expr):
-                facet_columns = facet.meta.root_names()
-                columns_to_select.extend(col for col in facet_columns if col not in columns_to_select)
-                facet_name = facet.meta.output_name()
-            else:
-                if facet not in columns_to_select:
-                    columns_to_select.append(facet)
-                facet_name = facet
-        else:
-            facet_name = None
-        df = (
-            (self.datamart.aggregates.last() if last else self.datamart._require_model_data())
-            .select(*columns_to_select)
-            .with_columns((pl.col("Performance") * pl.lit(100)).round(rounding))
-        )
-
-        if facet_name is not None and isinstance(facet, pl.Expr):
-            df = df.with_columns(facet.alias(facet_name))
-        df = cdh_utils._apply_query(df, query)
-
-        if return_df:
-            return df
-
-        import plotly.express as px
-
-        title = "over all models"
-        fig = px.scatter(
-            df.collect().with_columns(
-                pl.col("LastUpdate").dt.strftime("%v"),
-                pl.col(self.datamart.context_keys).fill_null(""),
-            ),
-            x="Performance",
-            y="SuccessRate",
-            color=color,
-            size="ResponseCount",
-            facet_col=facet_name,
-            facet_col_wrap=2,
-            hover_name="Name",
-            hover_data=["ModelID"] + self.datamart.context_keys + ["LastUpdate"],
-            title=f"Bubble Chart {title}",
-            template="pega",
-            facet_row_spacing=0.01,
-            labels={"LastUpdate": "Last Updated"},
-        )
-        fig = add_bottom_left_text_to_bubble_plot(fig, df, 1)
-        if show_metric_limits:
-            fig = add_metric_limit_lines(fig)
-        fig.update_traces(marker=dict(line=dict(color="black")))
-        fig.update_yaxes(tickformat=".3%")
-        return fig
-
-    @requires({"SnapshotTime"})
-    def over_time(
-        self,
-        metric: str = "Performance",
-        by: pl.Expr | str | list[str] = "ModelID",
-        *,
-        every: str | timedelta = "1d",
-        cumulative: bool = True,
-        query: QUERY | None = None,
-        facet: str | None = None,
-        show_metric_limits: bool = False,
-        return_df: bool = False,
-    ):
-        """Statistics over time
-
-        Parameters
-        ----------
-        metric : str, optional
-            The metric to plot, by default "Performance"
-        by : Union[pl.Expr, str, list[str]], optional
-            The column(s) to group by, by default "ModelID". When a list of
-            column names is passed, the values are concatenated with " / " into
-            a single combined dimension that is encoded as colour. To keep the
-            chart readable, the top 10 combinations by total ``metric`` are
-            kept and the rest are dropped with a warning.
-        every : Union[str, timedelta], optional
-            By what time period to group, by default "1d", see https://docs.pola.rs/api/python/stable/reference/expressions/api/polars.Expr.dt.truncate.html
-            for periods.
-        cumulative : bool, optional
-            Whether to show cumulative values or period-over-period changes, by default True
-        query : Optional[QUERY], optional
-            The query to apply to the data, by default None
-        facet : Optional[str], optional
-            Whether to facet the plot into subplots, by default None
-        show_metric_limits : bool, optional
-            Whether to show dashed horizontal lines at the metric limit
-            thresholds (from MetricLimits.csv), by default False.
-            Only applies when metric is "Performance".
-        return_df : bool, optional
-            Whether to return a dataframe instead of a plot, by default False
-
-        Returns
-        -------
-        Figure | pl.LazyFrame
-            Plotly line chart or LazyFrame if return_df=True
-
-        Examples
-        --------
-        >>> # Default: performance over time, one line per model
-        >>> fig = dm.plot.over_time()
-
-        >>> # SuccessRate over time, grouped by Channel
-        >>> fig = dm.plot.over_time(metric="SuccessRate", by="Channel")
-
-        >>> # Group by multiple dimensions at once (combined into a single
-        >>> # colour-encoded series)
-        >>> fig = dm.plot.over_time(by=["Channel", "Direction"])
-
-        >>> # Period-over-period response-count changes, faceted by Direction
-        >>> fig = dm.plot.over_time(
-        ...     metric="ResponseCount",
-        ...     cumulative=False,
-        ...     facet="Direction",
-        ... )
-
-        >>> # Return the data instead of the figure
-        >>> df = dm.plot.over_time(return_df=True)
-
-        """
-        percentage_metrics = ["Performance", "SuccessRate"]
-        metric_formatting = {
-            "SuccessRate": ":.4%",
-            "Performance": ":.2",
-            "Positives": ":.d",
-            "ResponseCount": ":.d",
-        }
-        if isinstance(by, list):
-            if len(by) == 0:
-                raise ValueError("`by` must contain at least one column name")
-            if len(by) == 1:
-                by = pl.col(by[0])
-            else:
-                by_name = " / ".join(by)
-                by = pl.concat_str(
-                    [pl.col(c).cast(pl.Utf8).fill_null("<null>") for c in by],
-                    separator=" / ",
-                ).alias(by_name)
-        elif not isinstance(by, pl.Expr):
-            by = pl.col(by)
-        by_col = by.meta.output_name()
-
-        if self.datamart.model_data is None:
-            raise ValueError("Visualisation requires model_data")
-
-        columns_to_select: set[str] = {"SnapshotTime", metric, "ResponseCount"}
-        columns_to_select.update(by.meta.root_names())
-        if facet:
-            columns_to_select.add(facet)
-
-        is_percentage = metric in percentage_metrics
-        metric_scaling = pl.lit(100.0 if metric == "Performance" else 1.0)
-
-        df = (
-            cdh_utils._apply_query(self.datamart.model_data, query).sort("SnapshotTime").select(list(columns_to_select))
-        )
-
-        grouping_columns = [by_col]
-        if facet:
-            grouping_columns.append(facet)
-        df = df.with_columns(by).set_sorted("SnapshotTime")
-
-        # Filter out null values in SnapshotTime and grouping columns to avoid issues with group_by_dynamic
-        null_filters = [pl.col("SnapshotTime").is_not_null()] + [pl.col(col).is_not_null() for col in grouping_columns]
-        df = df.filter(pl.all_horizontal(null_filters))
-
-        agg_expr = [
-            (
-                metric_scaling * cdh_utils.weighted_average_polars(metric, "ResponseCount")
-                if is_percentage
-                else pl.sum(metric)
-            ).alias(metric),
-        ]
-        if is_percentage:
-            agg_expr.append(pl.sum("ResponseCount"))
-
-        df = (
-            df.with_columns(pl.col("SnapshotTime"))
-            .group_by(grouping_columns + ["SnapshotTime"])
-            .agg(agg_expr)
-            .sort("SnapshotTime", by_col)
-        )
-
-        unique_intervals = df.select(pl.col("SnapshotTime").unique()).collect().height
-
-        if not cumulative and unique_intervals <= 1:
-            logger.warning(
-                f"Only one {every} interval of data found. Cannot calculate interval-to-interval differences. "
-                "Automatically switching to cumulative mode to show values instead.",
-            )
-            cumulative = True
-
-        plot_metric = metric
-        if not cumulative:
-            plot_metric = f"{metric}_change"
-            df = (
-                df.group_by_dynamic(
-                    "SnapshotTime",
-                    every=every,
-                    group_by=grouping_columns,
-                )
-                .agg(pl.last(metric))
-                .with_columns(
-                    pl.col(metric).diff().over(grouping_columns).alias(plot_metric),
-                )
-            )
-
-        if return_df:
-            return df
-        final_df = df.collect()
-
-        max_series = 10
-        unique_series = final_df.get_column(by_col).unique().to_list()
-        if len(unique_series) > max_series:
-            ranking_metric = "ResponseCount" if is_percentage else metric
-            top_series = (
-                final_df.group_by(by_col)
-                .agg(pl.sum(ranking_metric).alias("__total"))
-                .sort("__total", descending=True, nulls_last=True)
-                .head(max_series)
-                .get_column(by_col)
-                .to_list()
-            )
-            dropped = len(unique_series) - len(top_series)
-            logger.warning(
-                "over_time: %d unique values for `%s` exceeds the cap of %d; "
-                "keeping the top %d by total %s and dropping %d.",
-                len(unique_series),
-                by_col,
-                max_series,
-                max_series,
-                ranking_metric,
-                dropped,
-            )
-            final_df = final_df.filter(pl.col(by_col).is_in(top_series))
-
-        import plotly.express as px
-
-        facet_col_wrap = None
-        if facet:
-            unique_facet_values = final_df.select(facet).unique().height
-            facet_col_wrap = max(2, int(unique_facet_values**0.5))
-
-        title = "over all models" if facet is None else f"per {facet}"
-        fig = px.line(
-            final_df,
-            x="SnapshotTime",
-            y=plot_metric,
-            color=by_col,
-            hover_data={
-                by_col: ":.d",
-                plot_metric: metric_formatting[metric.split("_")[0]],
-            },
-            markers=True,
-            title=f"{metric} over time, per {by_col} {title}",
-            facet_col=facet,
-            facet_col_wrap=facet_col_wrap,
-            template="pega",
-        )
-
-        if metric == "SuccessRate":
-            fig.update_yaxes(tickformat=".2%")
-            fig.update_layout(yaxis={"rangemode": "tozero"})
-
-        if show_metric_limits and metric == "Performance":
-            fig = add_metric_limit_lines(fig, orientation="horizontal")
-
-        return fig
-
-    @requires({"ModelID", "Name"})
-    def proposition_success_rates(
-        self,
-        metric: str = "SuccessRate",
-        by: str = "Name",
-        *,
-        top_n: int = 0,
-        query: QUERY | None = None,
-        facet: str | None = None,
-        return_df: bool = False,
-    ):
-        """Proposition Success Rates
-
-        Parameters
-        ----------
-        metric : str, optional
-            The metric to plot, by default "SuccessRate"
-        by : str, optional
-            By which column to group the, by default "Name"
-        top_n : int, optional
-            Whether to take a top_n on the `by` column, by default 0
-        query : Optional[QUERY], optional
-            A query to apply to the data, by default None
-        facet : Optional[str], optional
-            What facetting column to apply to the graph, by default None
-        return_df : bool, optional
-            Whether to return a DataFrame instead of the graph, by default False
-
-        Returns
-        -------
-        Figure | pl.LazyFrame
-            Plotly histogram figure or LazyFrame if return_df=True
-
-        Examples
-        --------
-        >>> # Default: average success rate per proposition
-        >>> fig = dm.plot.proposition_success_rates()
-
-        >>> # Top-10 propositions by success rate, faceted by Channel
-        >>> fig = dm.plot.proposition_success_rates(top_n=10, facet="Channel")
-
-        >>> # Use ResponseCount as the metric grouped by Configuration
-        >>> fig = dm.plot.proposition_success_rates(
-        ...     metric="ResponseCount",
-        ...     by="Configuration",
-        ... )
-
-        """
-        if self.datamart.model_data is None:
-            raise ValueError("Visualisation requires model_data")
-
-        df = cdh_utils._apply_query(self.datamart.model_data, query).select(
-            {"ModelID", "Name", metric, by, facet},
-        )
-
-        if return_df:
-            return df
-
-        import plotly.express as px
-
-        title = "over all models" if facet is None else f"per {facet}"
-        fig = px.histogram(
-            df.collect(),
-            x=metric,
-            y=by,
-            color=by,
-            facet_col=facet,
-            facet_col_wrap=5,
-            histfunc="avg",
-            title=f"{metric} of each proposition {title}",
-            template="pega",
-        )
-        fig.update_yaxes(categoryorder="total ascending")
-        fig.update_layout(showlegend=False)
-        fig.update_yaxes(dtick=1, automargin=True)
-        errors = df.group_by(by).agg(pl.std("SuccessRate").fill_null(0)).collect()
-        for i, bar in enumerate(fig.data):
-            fig.data[i]["error_x"] = {
-                "array": [errors.filter(Name=bar["name"])["SuccessRate"].item()],
-                "valueminus": 0,
-            }
-        return fig
-
-    @requires(
-        combined_columns={
-            "PredictorName",
-            "Name",
-            "BinIndex",
-            "BinSymbol",
-            "BinResponseCount",
-            "BinPropensity",
-            "ModelID",
-        },
-    )
-    def score_distribution(
-        self,
-        model_id: str,
-        *,
-        active_range: bool = True,
-        return_df: bool = False,
-    ):
-        """Generate a score distribution plot for a specific model.
-
-        Parameters
-        ----------
-        model_id : str
-            The ID of the model to generate the score distribution for
-        active_range : bool, optional
-            Whether to filter to active score range only, by default True
-        return_df : bool, optional
-            Whether to return a dataframe instead of a plot, by default False
-
-        Returns
-        -------
-        Union[Figure, pl.LazyFrame]
-            Plotly figure showing score distribution or DataFrame if return_df=True
-
-        Raises
-        ------
-        ValueError
-            If no data is available for the provided model ID
-
-        Examples
-        --------
-        >>> # Score distribution for a specific model (active range only)
-        >>> fig = dm.plot.score_distribution(model_id="M-1001")
-
-        >>> # Include the full score range rather than just the active range
-        >>> fig = dm.plot.score_distribution(model_id="M-1001", active_range=False)
-
-        >>> # Retrieve the underlying binning data
-        >>> df = dm.plot.score_distribution(model_id="M-1001", return_df=True)
-
-        """
-        df = (
-            self.datamart.aggregates.last(table="combined_data")
-            .select(
-                {
-                    "PredictorName",
-                    "Name",
-                    "BinIndex",
-                    "BinSymbol",
-                    "BinResponseCount",
-                    "BinPropensity",
-                    "ModelID",
-                    "Configuration",
-                }
-                | set(
-                    self.datamart.context_keys,
-                ),
-            )
-            .filter(PredictorName="Classifier", ModelID=model_id)
-        ).sort("BinIndex")
-
-        if active_range:
-            active_ranges = self.datamart.active_ranges(model_id).collect()
-            if active_ranges.height > 0:
-                active_range_info = active_ranges.to_dicts()[0]
-                active_range_filter_expr = (pl.col("BinIndex") >= active_range_info["idx_min"]) & (
-                    pl.col("BinIndex") <= active_range_info["idx_max"]
-                )
-                df = df.filter(active_range_filter_expr)
-
-        if df.select(pl.first().len()).collect().item() == 0:
-            raise ValueError(f"There is no data for the provided modelid '{model_id}'")
-
-        if return_df:
-            return df
-
-        context = "/".join(
-            df.select(
-                pl.col("Configuration", *self.datamart.context_keys).fill_null(
-                    "MISSING",
-                ),
-            )
-            .unique()
-            .collect()
-            .row(0),
-        )
-        return distribution_graph(
-            df,
-            f"""Classifier score distribution<br>
-            <sup>{context}</sup>
-            """,
-        )
-
-    def multiple_score_distributions(
-        self,
-        query: QUERY | None = None,
-        show_all: bool = True,
-    ) -> list[Figure]:
-        """Generate the score distribution plot for all models in the query
-
-        Parameters
-        ----------
-        query : Optional[QUERY], optional
-            A query to apply to the data, by default None
-        show_all : bool, optional
-            Whether to 'show' all plots or just get a list of them, by default True
-
-        Returns
-        -------
-        list[go.Figure]
-            A list of Plotly charts, one for each model instance
-
-        Examples
-        --------
-        >>> # Generate and display score distributions for all models
-        >>> figs = dm.plot.multiple_score_distributions()
-
-        >>> # Collect figures without displaying them
-        >>> figs = dm.plot.multiple_score_distributions(show_all=False)
-
-        >>> # Limit to a specific channel before generating all distributions
-        >>> figs = dm.plot.multiple_score_distributions(
-        ...     query={"Channel": "Web"},
-        ...     show_all=False,
-        ... )
-
-        """
-        plots = []
-        for model_id in (
-            cdh_utils._apply_query(
-                self.datamart.aggregates.last(table="combined_data"),
-                query,
-            )
-            .select(pl.col("ModelID").unique())
-            .collect()["ModelID"]
-        ):
-            fig = self.score_distribution(model_id=model_id)
-            if show_all:
-                fig.show()
-            plots.append(fig)
-        return plots
-
-    @requires(
-        combined_columns={
-            "PredictorName",
-            "Name",
-            "ModelID",
-            "Configuration",
-            "BinIndex",
-            "BinSymbol",
-            "BinResponseCount",
-            "BinPropensity",
-        },
-    )
-    def predictor_binning(
-        self,
-        model_id: str,
-        predictor_name: str,
-        return_df: bool = False,
-    ):
-        """Generate a predictor binning plot for a specific model and predictor.
-
-        Parameters
-        ----------
-        model_id : str
-            The ID of the model containing the predictor
-        predictor_name : str
-            Name of the predictor to analyze
-        return_df : bool, optional
-            Whether to return a dataframe instead of a plot, by default False
-
-        Returns
-        -------
-        Union[Figure, pl.LazyFrame]
-            Plotly figure showing predictor binning or DataFrame if return_df=True
-
-        Raises
-        ------
-        ValueError
-            If no data is available for the provided model ID and predictor name
-
-        Examples
-        --------
-        >>> # Predictor binning for a single predictor on a model
-        >>> fig = dm.plot.predictor_binning(
-        ...     model_id="M-1001",
-        ...     predictor_name="Customer.Age",
-        ... )
-
-        >>> # Retrieve the underlying binning data
-        >>> df = dm.plot.predictor_binning(
-        ...     model_id="M-1001",
-        ...     predictor_name="Customer.Age",
-        ...     return_df=True,
-        ... )
-
-        """
-        df = (
-            self.datamart.aggregates.last(table="combined_data")
-            .select(
-                {
-                    "PredictorName",
-                    "Name",
-                    "ModelID",
-                    "Configuration",
-                    "BinIndex",
-                    "BinSymbol",
-                    "BinResponseCount",
-                    "BinPropensity",
-                }
-                | set(
-                    self.datamart.context_keys,
-                ),
-            )
-            .filter(
-                pl.col("PredictorName") == predictor_name,
-                pl.col("ModelID") == model_id,
-            )
-        ).sort("BinIndex")
-
-        if df.select(pl.first().len()).collect().item() == 0:
-            raise ValueError(
-                f"There is no data for the provided modelid {model_id} and predictor {predictor_name}",
-            )
-
-        if return_df:
-            return df
-        context = "/".join(
-            df.select(
-                pl.col("Configuration", *self.datamart.context_keys).fill_null(
-                    "MISSING",
-                ),
-            )
-            .unique()
-            .collect()
-            .row(0),
-        )
-        return distribution_graph(
-            df,
-            title=f"""Predictor binning for {predictor_name}<br><sup>{context}""",
-        )
-
-    def multiple_predictor_binning(
-        self,
-        model_id: str,
-        query: QUERY | None = None,
-        show_all=True,
-    ) -> list[Figure]:
-        """Generate predictor binning plots for all predictors in a model.
-
-        Parameters
-        ----------
-        model_id : str
-            The ID of the model to generate predictor binning plots for
-        query : Optional[QUERY], optional
-            A query to apply to the predictor data, by default None
-        show_all : bool, optional
-            Whether to display all plots or just return the list, by default True
-
-        Returns
-        -------
-        list[Figure]
-            A list of Plotly figures, one for each predictor in the model
-
-        Examples
-        --------
-        >>> # Plot binning for every predictor in a model
-        >>> figs = dm.plot.multiple_predictor_binning(model_id="M-1001")
-
-        >>> # Collect figures without displaying, filtered to IH predictors
-        >>> figs = dm.plot.multiple_predictor_binning(
-        ...     model_id="M-1001",
-        ...     query={"PredictorCategory": "IH"},
-        ...     show_all=False,
-        ... )
-
-        """
-        plots = []
-        for predictor in (
-            cdh_utils._apply_query(
-                self.datamart.aggregates.last(table="predictor_data"),
-                query,
-            )
-            .filter(pl.col("ModelID") == model_id)
-            .select(pl.col("PredictorName").unique())
-            .sort("PredictorName")
-            .collect()["PredictorName"]
-        ):
-            fig = self.predictor_binning(model_id=model_id, predictor_name=predictor)
-            if show_all:
-                fig.show()
-            plots.append(fig)
-        return plots
-
-    def _boxplot_pre_aggregated(
-        self,
-        df: pl.LazyFrame,
-        *,
-        y_col: str,
-        metric_col: str,
-        metric_weight_col: str | None = None,
-        legend_col: str | None = None,
-        color_discrete_map: dict[str, str] | None = None,
-        return_df: bool = False,
-    ):
-        if legend_col is None:
-            legend_col = y_col
-
-        if metric_weight_col is None:
-            mean_expr = pl.mean(metric_col)
-        else:
-            mean_expr = cdh_utils.weighted_average_polars(metric_col, metric_weight_col)
-        pre_aggs = (
-            df.group_by(list({y_col, legend_col}))
-            .agg(
-                median=pl.median(metric_col),
-                mean=mean_expr,
-                q1=pl.quantile(metric_col, 0.25),
-                q3=pl.quantile(metric_col, 0.75),
-                min=pl.min(metric_col),
-                max=pl.max(metric_col),
-                whisker_low=pl.quantile(metric_col, 0.1),
-                whisker_high=pl.quantile(metric_col, 0.9),
-            )
-            .with_columns(iqr=pl.col("q3") - pl.col("q1"))
-            # limiting the decimals should reduce the size of
-            # the plotly data in HTML significantly. TODO not sure thats true!!
-            .with_columns(cs.numeric().round(3))
-            .sort(["median", y_col], descending=True)
-            .collect()
-        )
-
-        if return_df:
-            return pre_aggs
-
-        if pre_aggs.select(pl.first().len()).item() == 0:
-            return None
-
-        import plotly.graph_objects as go
-        from ..utils.pega_template import colorway
-
-        fig = go.Figure()
-
-        # Build a color map from the provided global map, falling back to the
-        # pega colorway for any category not covered.
-        present_categories = sorted(pre_aggs.select(pl.col(legend_col).unique())[legend_col].to_list())
-        base_map: dict[str, str] = color_discrete_map or {}
-        fallback_index = 0
-        color_map: dict[str, str] = {}
-        for cat in present_categories:
-            if cat in base_map:
-                color_map[cat] = base_map[cat]
-            else:
-                color_map[cat] = colorway[fallback_index % len(colorway)]
-                fallback_index += 1
-
-        # Track which categories have been added to legend
-        legend_added = set()
-
-        # Create a box plot for each predictor
-        for i, row in enumerate(pre_aggs.iter_rows(named=True)):
-            show_in_legend = row[legend_col] not in legend_added
-            if show_in_legend:
-                legend_added.add(row[legend_col])
-
-            fig.add_trace(
-                go.Box(
-                    q1=[row["q1"]],
-                    median=[row["median"]],
-                    q3=[row["q3"]],
-                    lowerfence=[row["whisker_low"]],
-                    upperfence=[row["whisker_high"]],
-                    mean=[row["mean"]],
-                    y=[row[y_col]],
-                    boxpoints=False,
-                    marker=dict(color=color_map[row[legend_col]]),
-                    name=row[legend_col],
-                    legendgroup=row[legend_col],
-                    orientation="h",
-                    showlegend=show_in_legend,
-                ),
-            )
-
-        fig.update_layout(
-            title=f"{metric_col} by {legend_col}",
-            xaxis_title=metric_col,
-            yaxis_title="",
-            template="pega",
-        )
-
-        # set y-axis category order to show highest median values at the top
-        fig.update_yaxes(
-            categoryorder="array",
-            categoryarray=pre_aggs[y_col].to_list(),
-            automargin=True,
-            autorange="reversed",
-        )
-        return fig
-
-    @requires(
-        combined_columns={
-            "ModelID",
-            "PredictorName",
-            "PredictorCategory",
-            "ResponseCountBin",
-            "Type",
-            "EntryType",
-        },
-    )
-    def predictor_performance(
-        self,
-        *,
-        metric: str = "Performance",
-        top_n: int | None = None,
-        active_only: bool = False,
-        query: QUERY | None = None,
-        return_df: bool = False,
-    ):
-        """Plots a box plot of the performance of the predictors
-
-        Use the query argument to drill down to a more specific subset
-        If top n is given, chooses the top predictors based on the
-        weighted average performance across models, ordered by their median performance.
-
-        Parameters
-        ----------
-        metric : str, optional
-            The metric to plot, by default "Performance"
-            This is more for future-proofing, once FeatureImportance gets more used.
-        top_n : Optional[int], optional
-            The top n predictors to plot, by default None
-        active_only : bool, optional
-            Whether to only consider active predictor performance, by default False
-        query : Optional[QUERY], optional
-            The query to apply to the data, by default None
-        return_df : bool, optional
-            Whether to return a dataframe instead of a plot, by default False
-
-        Returns
-        -------
-        Figure | pl.LazyFrame | None
-            Plotly box plot figure, LazyFrame if return_df=True, or None if no data
-
-        See Also
-        --------
-        pdstools.adm.ADMDatamart.apply_predictor_categorization : how to override the out of the box predictor categorization
-
-        Examples
-        --------
-        >>> # Default: all predictors ranked by performance
-        >>> fig = dm.plot.predictor_performance()
-
-        >>> # Top-15 active predictors only
-        >>> fig = dm.plot.predictor_performance(top_n=15, active_only=True)
-
-        >>> # Filter to a specific channel and return the raw data
-        >>> df = dm.plot.predictor_performance(
-        ...     query={"Channel": "Web"},
-        ...     return_df=True,
-        ... )
-
-        """
-        # in combined_data, the performance metric is renamed to PredictorPerformance
-        metric = "PredictorPerformance" if metric == "Performance" else metric
-        df = cdh_utils._apply_query(
-            self.datamart.aggregates.last(table="combined_data")
-            .with_columns(
-                pl.col("PredictorPerformance") * 100.0,
-            )
-            .filter(pl.col("EntryType") != "Classifier")
-            .filter(pl.col("ResponseCountBin") > 0)
-            .unique(subset=["ModelID", "PredictorName"], keep="first"),
-            query=query,
-            allow_empty=True,
-        )
-        if active_only:
-            df = df.filter(pl.col("EntryType") == "Active")
-        if top_n:
-            df = self.datamart.aggregates._top_n(
-                df,
-                top_n,
-                metric,
-            )
-
-        fig = self._boxplot_pre_aggregated(
-            df,
-            y_col="PredictorName",
-            metric_col=metric,
-            # It is confusing, but ResponseCountBin is the correct total response
-            # count for a predictor. It is the sum of BinResponseCount. Thanks SK :)
-            metric_weight_col="ResponseCountBin",
-            legend_col="PredictorCategory",
-            color_discrete_map=self.datamart.predictor_category_color_map,
-            return_df=return_df,
-        )
-
-        if fig is not None and not return_df:
-            fig.update_xaxes(title="Performance")
-            fig.update_layout(title=f"{metric} by Predictor Category")
-
-        return fig
-
-    @requires(
-        combined_columns={
-            "ModelID",
-            "PredictorName",
-            "PredictorCategory",
-            "ResponseCountBin",
-            "Type",
-            "EntryType",
-        },
-    )
-    def predictor_category_performance(
-        self,
-        *,
-        metric: str = "Performance",
-        active_only: bool = False,
-        query: QUERY | None = None,
-        return_df: bool = False,
-    ):
-        """Plot the predictor category performance
-
-        Parameters
-        ----------
-        metric : str, optional
-            The metric to plot, by default "Performance"
-        active_only : bool, optional
-            Whether to only analyze active predictors, by default False
-        query : Optional[QUERY], optional
-            An optional query to apply, by default None
-        return_df : bool, optional
-            An optional flag to get the dataframe instead, by default False
-
-        Returns
-        -------
-        px.Figure
-            A Plotly figure
-
-        See Also
-        --------
-        pdstools.adm.ADMDatamart.apply_predictor_categorization : how to override the out of the box predictor categorization
-
-        Examples
-        --------
-        >>> # Default: performance box plot per predictor category
-        >>> fig = dm.plot.predictor_category_performance()
-
-        >>> # Active predictors only, filtered to a specific channel
-        >>> fig = dm.plot.predictor_category_performance(
-        ...     active_only=True,
-        ...     query={"Channel": "Web"},
-        ... )
-
-        >>> # Return underlying data for further analysis
-        >>> df = dm.plot.predictor_category_performance(return_df=True)
-
-        """
-        metric = "PredictorPerformance" if metric == "Performance" else metric
-
-        df = cdh_utils._apply_query(
-            self.datamart.aggregates.last(table="combined_data")
-            .with_columns(
-                pl.col("PredictorPerformance") * 100.0,
-            )
-            .filter(pl.col("EntryType") != "Classifier")
-            .filter(pl.col("ResponseCountBin") > 0),
-            query=query,
-        )
-        if active_only:
-            df = df.filter(pl.col("EntryType") == "Active")
-
-        fig = self._boxplot_pre_aggregated(
-            df,
-            y_col="PredictorCategory",
-            metric_col="PredictorPerformance",
-            legend_col="PredictorCategory",
-            color_discrete_map=self.datamart.predictor_category_color_map,
-            return_df=return_df,
-        )
-        if fig is not None and not return_df:
-            fig.update_xaxes(title="Performance")
-
-        return fig
-
-    @requires(
-        combined_columns={
-            "PredictorName",
-            "PredictorPerformance",
-            "BinResponseCount",
-            "PredictorCategory",
-        },
-    )
-    def predictor_contribution(
-        self,
-        *,
-        by: str = "Configuration",
-        query: QUERY | None = None,
-        return_df: bool = False,
-    ):
-        """Plots the predictor contribution for each configuration
-
-        Parameters
-        ----------
-        by : str, optional
-            By which column to plot the contribution, by default "Configuration"
-        query : Optional[QUERY], optional
-            An optional query to apply to the data, by default None
-        return_df : bool, optional
-            An optional flag to get a Dataframe instead, by default False
-
-        Returns
-        -------
-        px.Figure
-            A plotly figure
-
-        See Also
-        --------
-        pdstools.adm.ADMDatamart.apply_predictor_categorization : how to override the out of the box predictor categorization
-
-        Examples
-        --------
-        >>> # Default: contribution per Configuration
-        >>> fig = dm.plot.predictor_contribution()
-
-        >>> # Contribution grouped by Channel
-        >>> fig = dm.plot.predictor_contribution(by="Channel")
-
-        >>> # Return the contribution data for further processing
-        >>> df = dm.plot.predictor_contribution(return_df=True)
-
-        """
-        df = (
-            cdh_utils._apply_query(
-                self.datamart.aggregates.last(table="combined_data"),
-                query,
-            )
-            .filter(pl.col("PredictorName") != "Classifier")
-            # Huh?
-            .with_columns((pl.col("PredictorPerformance") - 0.5) * 2)
-            .group_by(by, "PredictorCategory")
-            .agg(
-                Performance=cdh_utils.weighted_average_polars(
-                    "PredictorPerformance",
-                    "BinResponseCount",
-                ),
-            )
-            .with_columns(
-                Contribution=(pl.col("Performance") / pl.sum("Performance").over(by)) * 100,
-            )
-            .sort("PredictorCategory")
-        )
-        if return_df:
-            return df
-
-        import plotly.express as px
-
-        fig = px.bar(
-            df.collect(),
-            x="Contribution",
-            y=by,
-            color="PredictorCategory",
-            orientation="h",
-            template="pega",
-            title="Contribution of different sources",
-        )
-        return fig
-
-    @requires(
-        combined_columns={
-            "PredictorName",
-            "Name",
-            "Performance",
-            "PredictorPerformance",
-            "ResponseCountBin",
-        },
-    )
-    def predictor_performance_heatmap(
-        self,
-        *,
-        top_predictors: int = 20,
-        top_groups: int | None = None,
-        by: str = "Name",
-        active_only: bool = False,
-        query: QUERY | None = None,
-        return_df: bool = False,
-    ):
-        """Generate a heatmap showing predictor performance across different groups.
-
-        Parameters
-        ----------
-        top_predictors : int, optional
-            Number of top-performing predictors to include, by default 20
-        top_groups : int, optional
-            Number of top groups to include, by default None (all groups)
-        by : str, optional
-            Column to group by for the heatmap, by default "Name"
-        active_only : bool, optional
-            Whether to only include active predictors, by default False
-        query : Optional[QUERY], optional
-            Optional query to filter the data, by default None
-        return_df : bool, optional
-            Whether to return a dataframe instead of a plot, by default False
-
-        Returns
-        -------
-        Union[Figure, pl.LazyFrame]
-            Plotly heatmap figure or DataFrame if return_df=True
-
-        Examples
-        --------
-        >>> # Default: top-20 predictors vs proposition (Name)
-        >>> fig = dm.plot.predictor_performance_heatmap()
-
-        >>> # Top-10 predictors across top-5 configurations, active only
-        >>> fig = dm.plot.predictor_performance_heatmap(
-        ...     top_predictors=10,
-        ...     top_groups=5,
-        ...     by="Configuration",
-        ...     active_only=True,
-        ... )
-
-        >>> # Return the pivot data for further processing
-        >>> df = dm.plot.predictor_performance_heatmap(return_df=True)
-
-        """
-        if isinstance(by, str):
-            by_name = by
-        else:
-            by = by.alias("Predictor")
-            by_name = "Predictor"
-
-        df = self.datamart.aggregates.predictor_performance_pivot(
-            query=query,
-            by=by,
-            top_predictors=top_predictors,
-            top_groups=top_groups,
-            active_only=active_only,
-        )
-
-        # Filter out rows with null values in the grouping column before transpose
-        # to avoid "Column with new names can't have null values" error
-        filtered_df = df.filter(pl.col(by_name).is_not_null()).collect()
-
-        # Check if dataframe is empty after filtering
-        if filtered_df.height == 0:
-            if return_df:
-                return pl.LazyFrame({by_name: []})
-            return None
-
-        collected = filtered_df.transpose(
-            include_header=True,
-            header_name=by_name,
-            column_names=by_name,
-        )
-
-        if return_df:
-            return collected.lazy()
-
-        import plotly.express as px
-
-        title = "over all models"
-        fig = px.imshow(
-            collected.select(pl.all().exclude(by_name)),
-            text_auto=".3f",
-            aspect="auto",
-            color_continuous_scale=get_colorscale("Performance"),
-            title=f"Top predictors {title}",
-            range_color=[0.5, 1],
-            y=collected[by_name],
-        )
-
-        fig.update_yaxes(dtick=1, automargin=True)
-        fig.update_xaxes(
-            dtick=1,
-        )
-        return fig
-
-    def gains_chart(
-        self,
-        value: str,
-        *,
-        index: str | None = None,
-        by: str | list[str] | None = None,
-        query: QUERY | None = None,
-        title: str | None = None,
-        return_df: bool = False,
-    ) -> Figure | pl.LazyFrame:
-        """Generate a gains chart showing cumulative distribution of a metric.
-
-        Creates a gains/lift chart to visualize model response skewness. Shows what
-        percentage of the total value (e.g., responses, positives) is driven by what
-        percentage of models. Useful for identifying if a small number of models
-        drive most of the volume.
-
-        Parameters
-        ----------
-        value : str
-            Column name containing the metric to compute gains for (e.g., "ResponseCount", "Positives")
-        index : str, optional
-            Column name to normalize by (e.g., population size). If None, uses model count.
-        by : str | list[str], optional
-            Column(s) to group by for separate gain curves (e.g., "Channel" or ["Channel", "Direction"])
-        query : QUERY, optional
-            Optional query to filter the data before computing gains
-        title : str, optional
-            Chart title. If None, uses "Gains Chart"
-        return_df : bool, default False
-            If True, return the gains data instead of the figure
-
-        Returns
-        -------
-        Figure | pl.LazyFrame
-            Plotly figure showing the gains chart, or LazyFrame if return_df=True
-
-        Examples
-        --------
-        >>> # Single gains curve for response count
-        >>> fig = datamart.plot.gains_chart(value="ResponseCount")
-
-        >>> # Gains curves by channel for positives
-        >>> fig = datamart.plot.gains_chart(
-        ...     value="Positives",
-        ...     by=["Channel", "Direction"],
-        ...     title="Cumulative Positives by Channel"
-        ... )
-        """
-        from ..utils import report_utils, cdh_utils
-
-        # Get the last snapshot of data
-        df = self.datamart.aggregates.last()
-
-        # Apply query if provided
-        if query is not None:
-            df = cdh_utils._apply_query(df, query)
-
-        # Calculate gains
-        gains_data = report_utils.gains_table(df, value=value, index=index, by=by)
-
-        if return_df:
-            return gains_data.lazy()
-
-        import plotly.express as px
-
-        # Create the plot
-        if by is None:
-            fig = px.area(
-                gains_data,
-                x="cum_x",
-                y="cum_y",
-                title=title or "Gains Chart",
-                template="pega",
-            )
-        else:
-            by_as_list = by if isinstance(by, list) else [by]
-            # Create a combined label for the legend
-            legend_col = gains_data.select(pl.concat_str(by_as_list, separator="/").alias("By"))["By"]
-
-            fig = px.line(
-                gains_data,
-                x="cum_x",
-                y="cum_y",
-                color=legend_col,
-                title=title or "Gains Chart",
-                template="pega",
-            )
-            fig = fig.update_layout(legend_title="/".join(by_as_list))
-
-        # Add diagonal reference line (represents perfect equality)
-        fig.add_shape(
-            type="line",
-            line=dict(color="grey", dash="dash"),
-            x0=0,
-            x1=1,
-            y0=0,
-            y1=1,
-        )
-
-        # Configure axes and layout
-        fig = (
-            fig.update_yaxes(scaleanchor="x", scaleratio=1)
-            .update_layout(
-                autosize=False,
-                width=400,
-                height=400,
-            )
-            .update_yaxes(constrain="domain", title="% of Responders")
-            .update_xaxes(tickformat=",.0%", constrain="domain", title="% of Population")
-            .update_yaxes(tickformat=",.0%")
-        )
-
-        return fig
-
-    def performance_volume_distribution(
-        self,
-        *,
-        by: str | list[str] | None = None,
-        query: QUERY | None = None,
-        bin_width: int = 3,
-        title: str | None = None,
-        return_df: bool = False,
-    ) -> Figure | pl.LazyFrame:
-        """Generate a performance vs volume distribution chart.
-
-        Shows how response volume is distributed across different model performance
-        ranges. Helps identify if volume is driven by high-performing or low-performing
-        models. Ideally, most volume should be in the 60-80 AUC range.
-
-        Parameters
-        ----------
-        by : str | list[str], optional
-            Column(s) to group by for separate curves (e.g., "Channel" or ["Channel", "Direction"])
-            If None, creates a single curve for all data
-        query : QUERY, optional
-            Optional query to filter the data before analysis
-        bin_width : int, default 3
-            Width of performance bins in AUC points (default creates bins of 3: 50-53, 53-56, etc.)
-        title : str, optional
-            Chart title. If None, uses "Performance vs Volume"
-        return_df : bool, default False
-            If True, return the binned data instead of the figure
-
-        Returns
-        -------
-        Figure | pl.LazyFrame
-            Plotly figure showing performance distribution, or LazyFrame if return_df=True
-
-        Notes
-        -----
-        Performance is binned from 50-100 using the specified bin_width. The chart shows
-        what percentage of responses fall into each performance bin, grouped by the `by`
-        parameter if provided.
-
-        Examples
-        --------
-        >>> # Single curve for all channels
-        >>> fig = datamart.plot.performance_volume_distribution()
-
-        >>> # Separate curves per channel
-        >>> fig = datamart.plot.performance_volume_distribution(
-        ...     by=["Channel", "Direction"],
-        ...     title="Performance Distribution by Channel"
-        ... )
-        """
-        from ..utils import cdh_utils
-
-        # Get model data (raises ValueError if not loaded)
-        df = self.datamart._require_model_data()
-
-        # Apply query if provided
-        if query is not None:
-            df = cdh_utils._apply_query(df, query)
-
-        # Determine grouping columns
-        group_cols: list[str] = []
-        if by is not None:
-            by_list = by if isinstance(by, list) else [by]
-            group_cols = by_list.copy()
-
-        # Create combined grouping column if needed
-        if by is not None:
-            by_list = by if isinstance(by, list) else [by]
-            df = df.with_columns(pl.concat_str(by_list, separator="/").alias("_GroupBy"))
-            group_cols = ["_GroupBy"]
-
-        # Bin performance and aggregate
-        df = (
-            df.with_columns(
-                (pl.col("Performance") * 100)
-                .cut(breaks=[p for p in range(50, 100, bin_width)], left_closed=True)
-                .alias("PerformanceBinned"),
-            )
-            .group_by(group_cols + ["PerformanceBinned"])
-            .agg(
-                pl.sum("ResponseCount"),
-                (pl.min("Performance") * 100).round(2).alias("break_label"),
-            )
-        )
-
-        # Calculate proportions within each group
-        if group_cols:
-            df = df.with_columns(
-                (pl.col("ResponseCount") / pl.col("ResponseCount").sum().over(group_cols[0])).alias("Proportion")
-            )
-        else:
-            df = df.with_columns((pl.col("ResponseCount") / pl.col("ResponseCount").sum()).alias("Proportion"))
-
-        collected_df: pl.DataFrame = df.sort(group_cols + ["PerformanceBinned", "Proportion"]).collect()
-
-        if return_df:
-            return collected_df.lazy()
-
-        import plotly.graph_objects as go
-
-        # Create the plot
-        fig = go.Figure()
-
-        if group_cols:
-            # Multiple curves
-            groups = collected_df[group_cols[0]].unique().sort()
-            for group in groups:
-                group_df = collected_df.filter(pl.col(group_cols[0]) == group)
-                fig.add_trace(
-                    go.Scatter(
-                        x=group_df["PerformanceBinned"],
-                        y=group_df["Proportion"],
-                        line_shape="spline",
-                        name=group,
-                    )
-                )
-        else:
-            # Single curve
-            fig.add_trace(
-                go.Scatter(
-                    x=collected_df["PerformanceBinned"],
-                    y=collected_df["Proportion"],
-                    line_shape="spline",
-                    name="All Channels",
-                )
-            )
-
-        # Configure layout
-        fig = (
-            fig.update_yaxes(tickformat=",.0%")
-            .update_xaxes(
-                type="category",
-                categoryorder="array",
-                categoryarray=collected_df["PerformanceBinned"].unique().sort().to_list(),
-            )
-            .update_layout(
-                template="pega",
-                title=title or "Performance vs Volume",
-                xaxis_title="Model Performance",
-                yaxis_title="Percentage of Responses",
-            )
-        )
-
-        # Apply legend color ordering
-        fig = cdh_utils.legend_color_order(fig)
-
-        return fig
-
-    def tree_map(
-        self,
-        metric: Literal[
-            "ResponseCount",
-            "Positives",
-            "Performance",
-            "SuccessRate",
-            "percentage_without_responses",
-        ] = "Performance",
-        *,
-        by: str = "Name",
-        query: QUERY | None = None,
-        return_df: bool = False,
-    ):
-        """Generate a tree map visualization showing hierarchical model metrics.
-
-        Parameters
-        ----------
-        metric : Literal["ResponseCount", "Positives", "Performance", "SuccessRate", "percentage_without_responses"], optional
-            The metric to visualize in the tree map, by default "Performance"
-        by : str, optional
-            Column to group by for the tree map hierarchy, by default "Name"
-        query : Optional[QUERY], optional
-            Optional query to filter the data, by default None
-        return_df : bool, optional
-            Whether to return a dataframe instead of a plot, by default False
-
-        Returns
-        -------
-        Union[Figure, pl.LazyFrame]
-            Plotly treemap figure or DataFrame if return_df=True
-
-        Examples
-        --------
-        >>> # Default: Performance tree map grouped by proposition name
-        >>> fig = dm.plot.tree_map()
-
-        >>> # ResponseCount tree map grouped by Channel
-        >>> fig = dm.plot.tree_map(metric="ResponseCount", by="Channel")
-
-        >>> # Return underlying summary data
-        >>> df = dm.plot.tree_map(return_df=True)
-
-        """
-        # TODO: clean up implementation a bit
-
-        group_by = self.datamart.context_keys[: self.datamart.context_keys.index(by)] if by != "ModelID" else [by]
-        df = self.datamart.aggregates.model_summary(by=by, query=query).select(
-            pl.col(group_by).cast(pl.Utf8).fill_null("Missing"),
-            pl.col("count").alias("Model Count"),
-            pl.col("Percentage_without_responses").alias(
-                "Percentage without responses",
-            ),
-            pl.col("ResponseCount_sum").alias("Total number of responses"),
-            pl.col("Weighted_success_rate").alias("Weighted average Success Rate"),
-            pl.col("Weighted_performance").alias("Weighted average Performance"),
-            pl.col("Positives_sum").alias("Total number of positives"),
-        )
-        if return_df:
-            return df
-
-        import plotly.express as px
-
-        context_keys = [px.Constant("All contexts")] + group_by
-
-        colorscale = get_colorscale(metric)
-
-        label_map = {
-            "ResponseCount": "Total number of responses",
-            "Positives": "Total number of positives",
-            "Performance": "Weighted average Performance",
-            "SuccessRate": "Weighted average Success Rate",
-            "percentage_without_responses": "Percentage without responses",
-        }
-
-        hover_data = {
-            "Model Count": ":.d",
-            "Percentage without responses": ":.0%",
-            "Total number of positives": ":.d",
-            "Total number of responses": ":.d",
-            "Weighted average Success Rate": ":.3%",
-            "Weighted average Performance": ":.2%",
-        }
-
-        fig = px.treemap(
-            df.collect(),
-            path=context_keys,
-            color=label_map.get(metric),
-            values="Model Count",
-            title=f"{label_map.get(metric)} by {by}",
-            hover_data=hover_data,
-            color_continuous_scale=colorscale,
-            range_color=[0.5, 1] if metric == "Performance" else None,
-        )
-
-        return fig
-
-    def predictor_count(
-        self,
-        *,
-        by: str | list[str] = ["EntryType", "Type"],
-        query: QUERY | None = None,
-        return_df: bool = False,
-    ):
-        """Generate a box plot showing the distribution of predictor counts by type.
-
-        Parameters
-        ----------
-        by : Union[str, list[str]], optional
-            Column(s) to group predictors by, by default ["EntryType", "Type"]
-        query : Optional[QUERY], optional
-            Optional query to filter the data, by default None
-        return_df : bool, optional
-            Whether to return a dataframe instead of a plot, by default False
-
-        Returns
-        -------
-        Union[Figure, pl.LazyFrame]
-            Plotly box plot figure or DataFrame if return_df=True
-
-        Examples
-        --------
-        >>> # Default: predictor count distribution by EntryType and Type
-        >>> fig = dm.plot.predictor_count()
-
-        >>> # Distribution by EntryType only
-        >>> fig = dm.plot.predictor_count(by="EntryType")
-
-        >>> # Return the raw counts
-        >>> df = dm.plot.predictor_count(return_df=True)
-
-        """
-        # Normalize to list[str] — if `by` is a bare string, treating it as
-        # Iterable[str] would iterate over characters, giving wrong group_by keys.
-        by_list: list[str] = [by] if isinstance(by, str) else by
-
-        df = cdh_utils._apply_query(
-            self.datamart.aggregates.last(table="combined_data"),
-            query,
-        ).filter(pl.col("EntryType") != "Classifier")
-
-        collected = df.group_by(["ModelID"] + by_list).agg(Count=pl.n_unique("PredictorName")).collect()
-
-        if len(by_list) > 1:
-            collected = pl.concat(
-                [
-                    collected,
-                    collected.group_by(["ModelID"] + by_list[1:])
-                    .agg(pl.col("Count").sum())
-                    .with_columns(pl.lit("Overall").alias(by_list[0])),
-                ],
-                how="diagonal_relaxed",
-            )
-        collected = collected.sort(by_list)
-
-        if return_df:
-            return collected
-
-        import plotly.express as px
-
-        fig = px.box(
-            collected.with_columns(_Type=pl.concat_str(reversed(by_list), separator=" / ")),
-            x="Count",
-            y="_Type",
-            color=by_list[0],
-            template="pega",
-        )
-
-        # Update title and x-axis label if we have a figure (not None and not a DataFrame)
-        if fig is not None and not return_df:
-            fig.update_layout(title="Predictors by Type")
-            fig.update_xaxes(title="Number of Predictors")
-            fig.update_layout(margin=dict(l=150))
-            fig.update_yaxes(title="")
-            # Reverse the legend order
-            fig.update_layout(legend=dict(traceorder="reversed"))
-
-        return fig
-
-    def binning_lift(
-        self,
-        model_id: str,
-        predictor_name: str,
-        *,
-        query: QUERY | None = None,
-        return_df: bool = False,
-    ):
-        """Generate a binning lift plot for a specific predictor showing propensity lift per bin.
-
-        Parameters
-        ----------
-        model_id : str
-            The ID of the model containing the predictor
-        predictor_name : str
-            Name of the predictor to analyze for lift
-        query : Optional[QUERY], optional
-            Optional query to filter the predictor data, by default None
-        return_df : bool, optional
-            Whether to return a dataframe instead of a plot, by default False
-
-        Returns
-        -------
-        Union[Figure, pl.LazyFrame]
-            Plotly bar chart showing binning lift or DataFrame if return_df=True
-
-        Examples
-        --------
-        >>> # Lift chart for a single predictor on a model
-        >>> fig = dm.plot.binning_lift(
-        ...     model_id="M-1001",
-        ...     predictor_name="Customer.Age",
-        ... )
-
-        >>> # Return the lift data for further analysis
-        >>> df = dm.plot.binning_lift(
-        ...     model_id="M-1001",
-        ...     predictor_name="Customer.Age",
-        ...     return_df=True,
-        ... )
-
-        """
-        df = cdh_utils._apply_query(
-            (
-                self.datamart.aggregates.last(table="predictor_data")
-                .filter(
-                    pl.col("PredictorName") == predictor_name,
-                    pl.col("ModelID") == model_id,
-                )
-                .sort("BinIndex")
-            ),
-            query,
-        ).select(
-            "PredictorName",
-            "BinIndex",
-            "BinPositives",
-            "BinNegatives",
-            "BinSymbol",
-        )
-        cols = df.collect_schema().names()
-
-        if "Lift" not in cols:
-            df = df.with_columns(
-                (cdh_utils.lift(pl.col("BinPositives"), pl.col("BinNegatives")) - 1.0).alias("Lift"),
-            )
-
-        shading_expr = pl.col("BinPositives") <= 5 if "BinPositives" in cols else pl.lit(False)
-
-        plot_df = df.with_columns(
-            pl.when((pl.col("Lift") >= 0.0) & shading_expr.not_())
-            .then(pl.lit("pos"))
-            .when((pl.col("Lift") >= 0.0) & shading_expr)
-            .then(pl.lit("pos_shaded"))
-            .when((pl.col("Lift") < 0.0) & shading_expr.not_())
-            .then(pl.lit("neg"))
-            .otherwise(pl.lit("neg_shaded"))
-            .alias("Direction"),
-            BinSymbolAbbreviated=abbreviate_label_expr("BinSymbol"),
-        ).sort(["PredictorName", "BinIndex"])
-
-        if return_df:
-            return plot_df
-
-        import plotly.express as px
-
-        fig = px.bar(
-            plot_df.collect(),
-            x="Lift",
-            y="BinSymbolAbbreviated",
-            color="Direction",
-            color_discrete_map=LIFT_DIRECTION_COLORS,
-            orientation="h",
-            title=f"Propensity Lift for {predictor_name}",
-            template="pega",
-            custom_data=["PredictorName", "BinSymbol"],
-            facet_col_wrap=3,
-            category_orders=plot_df.collect().to_dict(),
-        )
-        fig.update_traces(
-            hovertemplate="<br>".join(
-                [
-                    "<b>%{customdata[0]}</b>",
-                    "%{customdata[1]}",
-                    "<b>Lift: %{x:.2%}</b>",
-                ],
-            ),
-        )
-        fig.add_vline(x=0, line_color="black")
-
-        fig.update_layout(
-            showlegend=False,
-            hovermode="y",
-        )
-        fig.update_xaxes(title="", tickformat=",.2%")
-        fig.update_yaxes(
-            type="category",
-            categoryorder="array",
-            automargin=True,
-            title="",
-            dtick=1,  # show all bins
-            matches=None,  # allow independent y-labels if there are row facets
-        )
-        simplify_facet_titles(fig)
-        return fig
-
-    def action_overlap(
-        self,
-        group_col: str | list[str] | pl.Expr = "Channel",
-        overlap_col="Name",
-        *,
-        show_fraction=True,
-        query: QUERY | None = None,
-        return_df: bool = False,
-    ):
-        """Generate an overlap matrix heatmap showing shared actions across different groups.
-
-        Parameters
-        ----------
-        group_col : Union[str, list[str], pl.Expr], optional
-            Column(s) to group by for overlap analysis, by default "Channel"
-        overlap_col : str, optional
-            Column containing values to analyze for overlap, by default "Name"
-        show_fraction : bool, optional
-            Whether to show overlap as fraction or absolute count, by default True
-        query : Optional[QUERY], optional
-            Optional query to filter the data, by default None
-        return_df : bool, optional
-            Whether to return a dataframe instead of a plot, by default False
-
-        Returns
-        -------
-        Union[Figure, pl.LazyFrame]
-            Plotly heatmap showing action overlap or DataFrame if return_df=True
-
-        Examples
-        --------
-        >>> # Default: action overlap across Channels
-        >>> fig = dm.plot.action_overlap()
-
-        >>> # Overlap shown as absolute counts rather than fractions
-        >>> fig = dm.plot.action_overlap(show_fraction=False)
-
-        >>> # Overlap of configurations across Channel/Direction combinations
-        >>> fig = dm.plot.action_overlap(
-        ...     group_col=["Channel", "Direction"],
-        ...     overlap_col="Configuration",
-        ... )
-
-        >>> # Return the overlap matrix as a DataFrame
-        >>> df = dm.plot.action_overlap(return_df=True)
-
-        """
-        df = cdh_utils._apply_query(
-            self.datamart._require_model_data(),
-            query,
-        )
-
-        if isinstance(group_col, list):
-            group_col_name = "/".join(group_col)
-            df = df.with_columns(
-                pl.concat_str(*group_col, separator="/").alias(group_col_name),
-            )
-        elif isinstance(group_col, pl.Expr):
-            group_col_name = group_col.meta.output_name()
-            df = df.with_columns(group_col.alias(group_col_name))
-        else:
-            group_col_name = group_col
-
-        overlap_data = cdh_utils.overlap_matrix(
-            df.group_by(group_col_name).agg(pl.col(overlap_col).unique()).sort(group_col_name).collect(),
-            overlap_col,
-            by=group_col_name,
-            show_fraction=show_fraction,
-        )
-
-        if return_df:
-            return overlap_data
-
-        import plotly.express as px
-
-        plt = px.imshow(
-            overlap_data.drop(group_col_name),
-            text_auto=".1%" if show_fraction else ".d",
-            aspect="equal",
-            title=f"Overlap of {overlap_col}s",
-            x=overlap_data[group_col_name],
-            y=overlap_data[group_col_name],
-            template="pega",
-            labels=dict(
-                x=f"{group_col_name} on x",
-                y=f"{group_col_name} on y",
-                color="Overlap",
-            ),
-        )
-        plt.update_coloraxes(showscale=False)
-        return plt
-
-    def partitioned_plot(
-        self,
-        func: Callable,
-        facets: list[dict[str, str | None]],
-        show_plots: bool = True,
-        *args,
-        **kwargs,
-    ):
-        """Execute a plotting function across multiple faceted subsets of data.
-
-        This method applies a given plotting function to multiple filtered subsets of data,
-        where each subset is defined by the facet conditions. It's useful for generating
-        multiple plots with different filter conditions applied.
-
-        Parameters
-        ----------
-        func : Callable
-            The plotting function to execute for each facet
-        facets : list[dict[str, Optional[str]]]
-            list of dictionaries defining filter conditions for each facet
-        show_plots : bool, optional
-            Whether to display the plots as they are generated, by default True
-        *args : tuple
-            Additional positional arguments to pass to the plotting function
-        **kwargs : dict
-            Additional keyword arguments to pass to the plotting function
-
-        Returns
-        -------
-        list[Figure]
-            list of Plotly figures, one for each facet condition
-
-        Examples
-        --------
-        >>> # Bubble chart separately for Web and Email channels
-        >>> figs = dm.plot.partitioned_plot(
-        ...     dm.plot.bubble_chart,
-        ...     facets=[{"Channel": "Web"}, {"Channel": "Email"}],
-        ... )
-
-        >>> # Same but suppress auto-display
-        >>> figs = dm.plot.partitioned_plot(
-        ...     dm.plot.bubble_chart,
-        ...     facets=[{"Channel": "Web"}, {"Channel": "Email"}],
-        ...     show_plots=False,
-        ... )
-
-        """
-        figs = []
-        existing_query = kwargs.get("query")
-        for facet in facets:
-            combined_query = existing_query
-            for k, v in facet.items():
-                if v is None:
-                    new_query = pl.col(k).is_null()
-                else:
-                    new_query = pl.col(k).eq(v)
-                if combined_query is not None:
-                    combined_query = cdh_utils._combine_queries(
-                        combined_query,
-                        new_query,
-                    )
-                else:
-                    combined_query = new_query
-
-            kwargs["query"] = combined_query
-            fig = func(*args, **kwargs)
-            figs.append(fig)
-        if show_plots and fig is not None:
-            # fig.update_layout(title=title)
-            fig.show()
-
-        return figs
